refactor(api): move assessment-result model
