+++ api/lib/application/assessment-results/assessment-result-controller.js
@@ -1,4 +1,4 @@
-import { AssessmentResult } from '../../domain/models/AssessmentResult.js';
+import { AssessmentResult } from '../../../src/shared/domain/models/AssessmentResult.js';
 import { CompetenceMark } from '../../domain/models/CompetenceMark.js';
 import * as assessmentResultService from '../../domain/services/assessment-result-service.js';
