+++ api/lib/domain/events/handle-certification-scoring.js
@@ -1,4 +1,4 @@
-import { AssessmentResult } from '../models/AssessmentResult.js';
+import { AssessmentResult } from '../../../src/shared/domain/models/AssessmentResult.js';
 import { CertificationScoringCompleted } from './CertificationScoringCompleted.js';
 import { CompetenceMark } from '../models/CompetenceMark.js';
 import bluebird from 'bluebird';
