+++ api/lib/infrastructure/repositories/assessment-result-repository.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { knex } from '../../../db/knex-database-connection.js';
 import { MissingAssessmentId, AssessmentResultNotCreatedError } from '../../domain/errors.js';
 import { DomainTransaction } from '../DomainTransaction.js';
-import { AssessmentResult } from '../../domain/models/AssessmentResult.js';
+import { AssessmentResult } from '../../../src/shared/domain/models/AssessmentResult.js';
 import { CompetenceMark } from '../../domain/models/CompetenceMark.js';
 
 function _toDomain({ assessmentResultDTO, competencesMarksDTO }) {
