+++ api/tests/integration/infrastructure/repositories/certification-repository_test.js
@@ -1,7 +1,7 @@
 import { expect, databaseBuilder, catchErr, knex } from '../../../test-helper.js';
 import * as certificationRepository from '../../../../lib/infrastructure/repositories/certification-repository.js';
 import { CertificationCourseNotPublishableError } from '../../../../lib/domain/errors.js';
-import { status } from '../../../../lib/domain/models/AssessmentResult.js';
+import { status } from '../../../../src/shared/domain/models/AssessmentResult.js';
 
 describe('Integration | Repository | Certification', function () {
   describe('#publishCertificationCoursesBySessionId', function () {
