+++ api/tests/integration/infrastructure/repositories/jury-certification-summary-repository_test.js
@@ -8,7 +8,7 @@ import {
   ImpactfulSubcategories,
 } from '../../../../src/certification/shared/domain/models/CertificationIssueReportCategory.js';
 
-import { status as assessmentResultStatuses } from '../../../../lib/domain/models/AssessmentResult.js';
+import { status as assessmentResultStatuses } from '../../../../src/shared/domain/models/AssessmentResult.js';
 import * as juryCertificationSummaryRepository from '../../../../lib/infrastructure/repositories/jury-certification-summary-repository.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
