+++ api/tests/tooling/domain-builder/factory/build-certification-assessment-score-v3.js
@@ -1,5 +1,5 @@
 import { CertificationAssessmentScoreV3 } from '../../../../lib/domain/models/CertificationAssessmentScoreV3.js';
-import { status as CertificationStatus } from '../../../../lib/domain/models/AssessmentResult.js';
+import { status as CertificationStatus } from '../../../../src/shared/domain/models/AssessmentResult.js';
 
 const buildCertificationAssessmentScoreV3 = function ({ nbPix = 100, status = CertificationStatus.VALIDATED } = {}) {
   return new CertificationAssessmentScoreV3({
