+++ api/tests/unit/domain/models/FinalizedSession_test.js
@@ -1,7 +1,7 @@
 import { expect, domainBuilder } from '../../../test-helper.js';
 import { FinalizedSession } from '../../../../lib/domain/models/FinalizedSession.js';
 import { JuryCertificationSummary } from '../../../../lib/domain/read-models/JuryCertificationSummary.js';
-import { status as assessmentResultStatuses } from '../../../../lib/domain/models/AssessmentResult.js';
+import { status as assessmentResultStatuses } from '../../../../src/shared/domain/models/AssessmentResult.js';
 import { CertificationIssueReportCategory } from '../../../../src/certification/shared/domain/models/CertificationIssueReportCategory.js';
 
 describe('Unit | Domain | Models | FinalizedSession', function () {
